Separate model/service layer for users
